Chat: (finally) removed the testHelpers module.
