+++ src/chat/__tests__/cards/NotesList.test.tsx
@@ -1,13 +1,12 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
 
+import { currentUserMock, notesMock } from '../../../utils/test-helpers';
 import { renderWithRedux } from '../../../utils/test-utils';
 
 import withProviders from '../../../components/withProviders';
 import NotesList from '../../components/cards/NotesList';
 import { initialEditMode } from '../../redux';
-
-import { currentUserMock, notesMock } from '../testHelpers';
 
 describe.skip('<NotesList />', () => {
   const NotesListWithUser = () => <NotesList notes={notesMock} />;
